bots228/max_web_app/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
jinja2==3.1.5
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.4
